refactor(utils): Move cache utilities to cache/ subdirectory
Phase 2 (Day 2): Cache utilities
- Moved 3 cache files to src/utils/cache/
- Updated 3 source imports + 3 test imports
- Fixed relative paths in cache-keys.ts and csv-processor-core.ts
- Tests passing (199/199 smoke tests)

Files moved:
- cache.ts
- cache-keys.ts
- kv-results-handler.ts
