+++ src/handlers/book-search.ts
@@ -13,7 +13,7 @@ import { UnifiedCacheService } from '../services/unified-cache'
 import type { Env } from '../types/env'
 import { writeCacheMetrics } from '../utils/analytics/analytics'
 import { detectImageQuality } from '../utils/book-metadata'
-import { setCached } from '../utils/cache'
+import { setCached } from '../utils/cache/cache'
 import { transformWorkToGoogleFormat } from '../utils/transform-work'
 
 // ============================================================================
