+++ src/utils/jobs/csv-processor-core.ts
@@ -14,7 +14,7 @@
 import { buildCSVParserPrompt } from '../../prompts/csv-parser-prompt'
 import { parseCSVWithGemini as parseCSVWithGeminiImpl } from '../../providers/gemini-csv-provider'
 import type { Env } from '../../types/env'
-import { generateCSVCacheKey } from '../cache-keys'
+import { generateCSVCacheKey } from '../cache/cache-keys'
 import { validateCSV as validateCSVImpl } from '../validation/csv-validator'
 import type { ProgressReporter } from './progress-reporter'
